Testing Cassandra projection
