+++ sourcing-new/src/main/scala/ch/epfl/bluebrain/nexus/sourcingnew/projections/ProjectionModule.scala
@@ -4,9 +4,8 @@ package ch.epfl.bluebrain.nexus.sourcingnew.projections
 import akka.actor.typed.ActorSystem
 import akka.stream.alpakka.cassandra.scaladsl.CassandraSession
 import cats.effect.{Async, ContextShift}
-import ch.epfl.bluebrain.nexus.sourcingnew.projections.cassandra.{CassandraProjection, CassandraSchemaManager}
+import ch.epfl.bluebrain.nexus.sourcingnew.projections.cassandra.{CassandraProjection, CassandraSchemaManager, ProjectionConfig, CassandraConfig}
 import ch.epfl.bluebrain.nexus.sourcingnew.projections.jdbc.{JdbcConfig, JdbcProjection, JdbcSchemaManager}
-import com.typesafe.config.Config
 import distage.{Axis, ModuleDef, Tag, TagK}
 import doobie.util.transactor.Transactor
 import io.circe.{Decoder, Encoder}
@@ -20,13 +19,13 @@ final class ProjectionModule[F[_]: ContextShift: Async: TagK, A: Encoder: Decode
 
   // Cassandra
   make[SchemaManager[F]].tagged(Persistence.Cassandra).from {
-    (cassandraSession: CassandraSession, journalConfig: Config, actorSystem: ActorSystem[Nothing]) =>
-      new CassandraSchemaManager[F](cassandraSession, journalConfig, actorSystem)
+    (cassandraSession: CassandraSession, schemaConfig: CassandraConfig, actorSystem: ActorSystem[Nothing]) =>
+      new CassandraSchemaManager[F](cassandraSession, schemaConfig, actorSystem)
 
   }
   make[Projection[F, A]].tagged(Persistence.Cassandra).from {
-    (cassandraSession: CassandraSession, journalConfig: Config, actorSystem: ActorSystem[Nothing]) =>
-      new CassandraProjection[F, A](cassandraSession, journalConfig, actorSystem)
+    (cassandraSession: CassandraSession, projectionConfig: ProjectionConfig, actorSystem: ActorSystem[Nothing]) =>
+      new CassandraProjection[F, A](cassandraSession, projectionConfig, actorSystem)
 
   }
