+++ sourcing-new/src/test/resources/cassandra.conf
@@ -1,0 +1,33 @@
+datastax-java-driver {
+  basic {
+    load-balancing-policy.local-datacenter = datacenter1
+  }
+  advanced.reconnect-on-init = true
+}
+akka {
+  actor.warn-about-java-serializer-usage = false
+  log-dead-letters-during-shutdown = true
+
+  persistence {
+    journal.plugin = "akka.persistence.cassandra.journal"
+    snapshot-store.plugin = "akka.persistence.cassandra.snapshot"
+
+    cassandra {
+      journal {
+        keyspace-autocreate = true
+        tables-autocreate = true
+        event-adapters {
+          tagging = "ch.epfl.bluebrain.nexus.sourcing.projections.Fixture$TaggingAdapter"
+        }
+        event-adapter-bindings {
+          "ch.epfl.bluebrain.nexus.sourcing.projections.Fixture$Event" = tagging
+        }
+      }
+
+      query {
+        refresh-interval = 50ms
+      }
+    }
+  }
+}
+
