+++ testkit/src/main/scala/ch/epfl/bluebrain/nexus/testkit/cassandra/CassandraDocker.scala
@@ -1,8 +1,12 @@
 package ch.epfl.bluebrain.nexus.testkit.cassandra
 
-import distage.{ModuleDef, TagK}
-import izumi.distage.docker.ContainerDef
+import cats.effect.{ConcurrentEffect, ContextShift, IO}
+import ch.epfl.bluebrain.nexus.testkit.DockerSupport
+import distage.TagK
 import izumi.distage.docker.Docker.DockerPort
+import izumi.distage.docker.modules.DockerSupportModule
+import izumi.distage.docker.{ContainerDef, Docker}
+import izumi.distage.model.definition.ModuleDef
 
 object CassandraDocker extends ContainerDef {
   val primaryPort: DockerPort = DockerPort.TCP(9042)
@@ -18,6 +22,26 @@ object CassandraDocker extends ContainerDef {
       )
     )
   }
+
+  class Module[F[_]: ConcurrentEffect: ContextShift: TagK] extends ModuleDef {
+    make[CassandraDocker.Container].fromResource {
+      CassandraDocker.make[F]
+    }
+
+    make[CassandraHostConfig].from { docker: CassandraDocker.Container =>
+      val knownAddress = docker.availablePorts.availablePorts(primaryPort).head
+      CassandraHostConfig(knownAddress.hostV4, knownAddress.port)
+    }
+
+    // add docker dependencies and override default configuration
+    include(new DockerSupportModule[IO] overridenBy new ModuleDef {
+      make[Docker.ClientConfig].from {
+        DockerSupport.clientConfig
+      }
+    })
+  }
+
+  final case class CassandraHostConfig(host: String, port: Int)
 }
 
 class CassandraDockerModule[F[_]: TagK] extends ModuleDef {
